Get rid of a compiler warning.
git-svn-id: svn://vcs.exim.org/pcre/code/trunk@1403 2f5784b3-3f2a-0410-8824-cb99058d5e15

+++ pcretest.c
@@ -3687,7 +3687,7 @@ while (!done)
               ppp = pp;
               while (*ppp != '>') ppp++;
               printf("** The %.*s modifier is locked out - giving up.\n", 
-                ppp - pp + 1, pp);
+                (int)(ppp - pp + 1), pp);
               yield = 1;
               goto EXIT;     
               }
